refactor: split addWorkoutPlanDialog into custom hook
