+++ SparkyFitnessFrontend/src/constants/exercises.ts
@@ -45,3 +45,13 @@ export const EXERCISE_CATEGORIES = [
     defaultLabel: 'Stretching',
   },
 ] as const;
+
+export const DAYS_OF_WEEK = [
+  { id: 0, name: 'Sunday' },
+  { id: 1, name: 'Monday' },
+  { id: 2, name: 'Tuesday' },
+  { id: 3, name: 'Wednesday' },
+  { id: 4, name: 'Thursday' },
+  { id: 5, name: 'Friday' },
+  { id: 6, name: 'Saturday' },
+];
